fix: first refactoring
